requirements.txt
setuptools>=69.0.0
wheel>=0.42.0
python-dotenv==1.0.0
openai==1.12.0
SpeechRecognition==3.10.1
pyttsx3==2.90
selenium==4.18.1
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
pywhatkit==5.4
imap-tools==1.10.0
pyautogui==0.9.54
keyboard==0.13.5
cryptography==42.0.5
pyinstaller==6.12.0
Pillow>=10.0.0 